In progress

+++ src/kernel/utils/vector.c
@@ -6,11 +6,11 @@ Vector vector_make() {
     return (Vector){ NULL, 0 };
 }
 
-//Status vector_resize(Vector* vector, size_t new_size, size_t sizeof_element) {
+//Status vector_resize(Vector* vector, size_t new_size, uint32_t sizeof_element) {
 //    return KERNEL_OK; 
 //}
 
-Status vector_push_back(Vector* vector, const void* src, size_t sizeof_element) {
+Status vector_push_back(Vector* vector, const void* src, uint32_t sizeof_element) {
     void* new_buffer = kmalloc((vector->size + 1) * sizeof_element);
 
     if (new_buffer == NULL) return KERNEL_ERROR;
@@ -33,11 +33,11 @@ Status vector_push_back(Vector* vector, const void* src, size_t sizeof_element) 
     return KERNEL_OK;
 }
 
-//void vector_pop_back(Vector* vector, size_t sizeof_element) {
+//void vector_pop_back(Vector* vector, uint32_t sizeof_element) {
 //    
 //}
 
-void vector_remove(Vector* vector, size_t idx, size_t sizeof_element) {
+void vector_remove(Vector* vector, size_t idx, uint32_t sizeof_element) {
 }
 
 void vector_clear(Vector* vector) {
